requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via requests
contourpy==1.0.7
    # via matplotlib
cryptography==40.0.2
    # via pyjwt
cycler==0.11.0
    # via matplotlib
deep-sort-realtime==1.3.2
    # via -r requirements.in
deprecated==1.2.13
    # via pygithub
fonttools==4.39.4
    # via matplotlib
idna==3.4
    # via requests
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.7.1
    # via supervision
numpy==1.24.2
    # via
    #   -r requirements.in
    #   contourpy
    #   deep-sort-realtime
    #   matplotlib
    #   opencv-python
    #   scipy
    #   supervision
opencv-python==4.7.0.72
    # via
    #   -r requirements.in
    #   deep-sort-realtime
    #   supervision
packaging==23.1
    # via matplotlib
pillow==9.5.0
    # via matplotlib
pycparser==2.21
    # via cffi
pygithub==1.58.2
    # via -r requirements.in
pyjwt[crypto]==2.7.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via matplotlib
requests==2.30.0
    # via pygithub
scipy==1.10.1
    # via deep-sort-realtime
six==1.16.0
    # via python-dateutil
supervision==0.3.1
    # via -r requirements.in
torch==1.13.1
    # via -r requirements.in
typing-extensions==4.5.0
    # via torch
urllib3==2.0.2
    # via requests
wrapt==1.15.0
    # via deprecated
